chore(cloud): separate lifecycle support types
