+++ Sources/Cloud/CloudTuiCreationFailure.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+/// Typed outcomes for a durable Cloud terminal creation intent.
+enum CloudTuiCreationFailure: Error, Equatable, Sendable {
+    case outcomeUnknown
+    case unsupported
+}
